+++ src/components/Form/UserLoginForm.jsx
@@ -4,11 +4,11 @@ import { ButtonPupple } from "../Share/ButtonPupple";
 import { UserFormGroup } from "../Login/LoginLayout";
 import { useNavigate } from "react-router-dom";
 
-// import axios from "axios";
-
-// axios.defaults.withCredentials = true;
+import axios from "axios";
 
 export default function UserInputForm({ onSubmit }) {
+    axios.defaults.withCredentials = true;
+
     let navigate = useNavigate();
     const [usrInputs, setUsrInputs] = useState({
         id: "",
@@ -33,32 +33,50 @@ export default function UserInputForm({ onSubmit }) {
             password: password,
         };
 
-        async function getUser() {
-            try {
-                const response = await fetch(
-                    "http://54.166.132.169:8080/api/user/login",
-                    {
-                        method: "POST",
-                        headers: { "Content-type": "application/json" },
-                        body: JSON.stringify(myData),
-                    }
-                );
+        axios
+            .post("http://54.166.132.169:8080/api/user/login", myData, {
+                withCredentials: true,
+                credentials: "include",
+            })
+            .then(function (response) {
+                console.log("성공", response);
+                alert("로그인 성공하였습니다.");
+            })
+            .catch(function (error) {
+                // 오류발생시 실행
+                console.log("실패", error);
+            });
 
-                if (!response.ok) {
-                    console.log("실패", response.status);
-                    throw new Error(`Error! status: ${response.status}`);
-                }
+        // async function getUser() {
+        //     try {
+        //         const response = await fetch(
+        //             "http://54.166.132.169:8080/api/user/login",
+        //             {
+        //                 method: "POST",
+        //                 headers: {
+        //                     "Content-Type": "application/x-www-form-urlencoded",
+        //                 },
+        //                 withCredentials: true,
+        //                 credentials: "include",
 
-                const result = await response.json();
-                console.log(result);
-                navigate("/exersise");
-                return result;
-            } catch (err) {
-                console.log(err);
-            }
-        }
+        //                 body: JSON.stringify(myData),
+        //             }
+        //         );
 
-        getUser();
+        //         if (!response.ok) {
+        //             throw new Error(`Error! status: ${response.status}`);
+        //         }
+
+        //         const result = await response.json();
+        //         console.log(result);
+        //         navigate("/exersise");
+        //         return result;
+        //     } catch (err) {
+        //         console.log(err);
+        //     }
+        // }
+
+        // getUser();
 
         // fetch("http://54.166.132.169:8080/api/user/login", {
         //     method: "POST",
